requirements.txt
streamlit
pandas
matplotlib
mlflow
scikit-learn
numpy
statsmodels
plotly.express
tensorflow
Pillow
joblib